+++ packages/connector/src/index.ts
@@ -1,5 +1,6 @@
 import { MetamaskSnap } from './snap';
-import { isMetamaskInstalled, isSnapInstalled } from './utils';
+import { isMetamaskInstalled } from './utils';
+import versionJson from './version.json';
 
 export { MetamaskSnap } from './snap';
 export * from './types';
@@ -9,7 +10,7 @@ export const defaultSnapOrigin = 'npm:@rarimo/rarime';
 
 export const enableSnap = async (
   snapOrigin?: string,
-  version = '0.4.x',
+  version = versionJson.version,
 ): Promise<MetamaskSnap> => {
   const snapId = snapOrigin ?? defaultSnapOrigin;
 
@@ -17,16 +18,12 @@ export const enableSnap = async (
     throw new Error('Metamask is not installed');
   }
 
-  const isInstalled = await isSnapInstalled(snapId, version);
-
-  if (!isInstalled) {
-    await window.ethereum.request({
-      method: 'wallet_requestSnaps',
-      params: {
-        [snapId]: { ...(version && { version }) },
-      },
-    });
-  }
+  await window.ethereum.request({
+    method: 'wallet_requestSnaps',
+    params: {
+      [snapId]: { ...(version && { version }) },
+    },
+  });
 
   return new MetamaskSnap(snapId);
 };