+++ packages/snap/src/helpers/common-helpers.ts
@@ -12,7 +12,9 @@ export const getChainInfo = (chainId: number): ChainInfo => {
 };
 
 export const getProviderChainInfo = async (): Promise<ChainInfo> => {
-  const provider = new providers.Web3Provider(window.ethereum);
+  const provider = new providers.Web3Provider(
+    (ethereum as any) as providers.ExternalProvider,
+  );
   const network = await provider.getNetwork();
 
   return getChainInfo(network.chainId);
@@ -30,4 +32,8 @@ export const getRarimoStateContractAddress = (chainId: number) => {
   return config.RARIMO_STATE_CONTRACT_ADDRESS[
     SUPPORTED_CHAINS[chainId].rarimoNetworkType
   ];
+};
+
+export const getHostname = (origin: string): string => {
+  return new URL(origin).hostname;
 };